Guards for student and teacher

+++ Sqorer/src/app/components/header-profile-popover/header-profile-popover.component.ts
@@ -1,6 +1,7 @@
+import { AuthService } from 'src/app/services/auth.service';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { PopoverController } from '@ionic/angular';
+import { PopoverController, AlertController } from '@ionic/angular';
 
 @Component({
   selector: 'app-header-profile-popover',
@@ -10,14 +11,23 @@ import { PopoverController } from '@ionic/angular';
 export class HeaderProfilePopoverComponent implements OnInit {
 
 
-  constructor(private popoverController: PopoverController, private router: Router) { }
+  constructor(private popoverController: PopoverController,
+    private router: Router, private authService: AuthService,
+    private alertController: AlertController) { }
 
   ngOnInit() { }
 
   logout() {
-    localStorage.clear();
-    this.popoverController.dismiss();
-    this.router.navigate(['/home']);
+
+    this.authService.logout().subscribe((data) => {
+      if (data.success) {
+        localStorage.clear();
+        this.popoverController.dismiss();
+        this.router.navigate(['/home']);
+      }
+    }, (err) => {
+      this.alert('Error', err.error.msg);
+    });
   }
 
   goProfile() {
@@ -25,6 +35,15 @@ export class HeaderProfilePopoverComponent implements OnInit {
     this.router.navigate(['/dashboard/profile']);
   }
 
-
+  async alert(header, msg) {
+    const alert = await this.alertController.create({
+      cssClass: 'my-custom-class',
+      header,
+      subHeader: '',
+      message: msg,
+      buttons: ['OK']
+    });
+    await alert.present();
+  }
 
 }
